% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_version_number.R
\name{as_version_number}
\alias{as_version_number}
\title{Coerce to a version number}
\usage{
as_version_number(x)
}
\arguments{
\item{x}{The object to coerce}
}
\value{
A version number
}
\description{
By default this just calls \code{version_number(x)}
}
\examples{

as_version_number("1.0.3.1")
}
