% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bump_version.R
\name{bump_version}
\alias{bump_version}
\title{Bump a Version Number}
\usage{
bump_version(x, which = getOption("versionr.parts"), n = 1)
}
\arguments{
\item{x}{Version numbers to bump}

\item{which}{Either a number indicating the part to bump or one of the
parts as specified in \code{getOption("versionr.parts")} - defaults are
"major", "minor", "patch" and "dev"}

\item{n}{The amount to increment the version number part by - will almost
always be 1}
}
\value{
A list of version numbers
}
\description{
Bump a Version Number
}
\examples{
nums <- version_number("1.0.0.0", "2.0")
bump_version(nums)

}
