% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rough_equality.R
\name{\%==\%}
\alias{\%==\%}
\title{Check for rough equality between version numbers}
\usage{
x \%==\% y
}
\arguments{
\item{x, y}{Version numbers}
}
\value{
Logical
}
\description{
Returns a logical vector indicating whether all the present version
parts match for two sets of version numbers
}
