% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.R
\name{latest}
\alias{latest}
\title{latest}
\usage{
latest(x)
}
\arguments{
\item{x}{The name of an R object in quotes}
}
\value{
The R object is created as a variable in the Global Environment, or replaces the object with the same name.
}
\description{
Restore the latest saved version of an object by name (in quotes) from the subdirectory of your working directory at ../Env/Versions/.
}
\examples{
latest("df")
}
