% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_version_number}
\alias{pad_version_number}
\title{Pad version number(s) with zeros}
\usage{
pad_version_number(x, places, warning = NULL, use_na = FALSE)
}
\arguments{
\item{x}{Version numbers}

\item{places}{The number of parts each number should have}

\item{warning}{Warning to use - NULL for no warning}

\item{use_na}{Whether to pad with NA instead of zeros}
}
\description{
Pad version number(s) with zeros
}
