% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_all_vars.R
\name{restore_all_vars}
\alias{restore_all_vars}
\title{restore_all_vars}
\usage{
restore_all_vars()
}
\value{
Backups saved in your working directory at ../Env/Versions/ are restored as objects to your Global Environment. Any objects in the Global Environment with the same variable name will be replaced/overwritten so backups should be made if needed.
}
\description{
Restores the latest timestamped version of any R object in the save_all_vars subdirectory. To avoid restoring objects you no longer need, manually delete them from the subdirectory before using this function.
}
\examples{
restore_all_vars()

}
