% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_all_vars.R
\name{save_all_vars}
\alias{save_all_vars}
\title{save_all_vars}
\usage{
save_all_vars(envir = .GlobalEnv)
}
\value{
Backups of all variables currently in the Global Environment are saved as .rds files within a subdirectory of your working directory at ../Env/Versions/.
}
\description{
Save all objects in the Global Environment
}
\examples{
save_all_vars()

}
