% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveVersion.R
\name{saveVersion}
\alias{saveVersion}
\title{saveVersion}
\usage{
saveVersion(x)
}
\arguments{
\item{x}{The name of an R object in quotes}
}
\value{
The R object is saved as a .rds file within a subdirectory of your working directory at ../Env/Versions/.
}
\description{
save an object by name (in quotes).
}
\examples{
saveVersion("df")
}
