% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_versions.R
\name{tabulate_versions}
\alias{tabulate_versions}
\title{Tabulate version numbers}
\usage{
tabulate_versions(x, full_version = c("none", "version_number", "character"))
}
\arguments{
\item{x}{The version numbers to tabulate}

\item{full_version}{Control how a column is added that contains the full
version numbers. Either "version_number" for a \code{version_number}-class column,
"character" for a \code{character}-class column or "none" to not add a column.}
}
\value{
A \code{data.frame}
}
\description{
This function expands the parts of a vector of version numbers into
a \code{data.frame} which may be useful, e.g. for moving version numbers
to a database.
}
\examples{
tabulate_versions(version_number("1.0.0.0", "1.0.1.0"))

tabulate_versions(version_number("1.0.0.0", "1.0.1.0"), "version_number")

}
