% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versionR.R
\name{saveVersion}
\alias{saveVersion}
\title{saveVersion() and latest()}
\usage{
saveVersion(x)
}
\arguments{
\item{"x"}{The name of an R object in quotes}
}
\value{
The R object is created as a variable in the Global Environment, or replaces the object with the same name.
}
\description{
versionR
}
\details{
Functions for saving and restoring timestamped versions of R objects within a subfolder of the working directory.
}
\section{Functions}{
\itemize{
\item \code{saveVersion}: two functions saveVersion() and latest() are used to save an object by name (in quotes) and restore the latest version of it. Subdirectories are created within your working directory > Env > Versions to store the R object as a .rds file.
}}

\section{Save R objects by variable name.}{
NA
}

\examples{
saveVersion("df")
latest("df")

}
