% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs.R
\name{version_number}
\alias{version_number}
\title{Create a version number}
\usage{
version_number(..., .list)
}
\arguments{
\item{...}{Can be either vectors of numbers, character strings or a single
character vector}

\item{.list}{Optionally supply the arguments as a list. Useful for
programming}
}
\description{
Create a version number
}
\examples{
version_number("1.2.3.4", c(1, 2, 3, 4))
version_number(.list = list("1.0.0", c(1, 0, 0)))

}
