% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_part.R
\name{version_part}
\alias{version_part}
\title{Get a Version Number Part}
\usage{
version_part(x, which = getOption("version.parts"))
}
\arguments{
\item{x}{Version numbers}

\item{which}{The part to get - either as a string or a number}
}
\value{
An integer vector
}
\description{
This is useful for checking the value of a certain part of a version
number, and is particularly useful when filtering datasets with a
version number column. If a part is not present, \code{NA} will be returned.
}
\examples{
nums <- version_number("1.0.0", "1.0.0.1", "1.0.0.2", "1.2.0", "2.0.0")

# Specify the part to get using a string or a number
version_part(nums, "dev")
version_part(nums, 4)
version_part(nums, "major")

# Filter datasets
if (require(dplyr, quietly = TRUE)) {

  version_desc <- c("It's alive!", "Tweak to balancing", "Bug fix",
                    "Can do flips", "Complete rework")

  # Get the numbers as a data.frame
  data <- tibble(versions = nums, version_desc = version_desc)
  data

}

# Filter for dev versions
if (require(dplyr, quietly = TRUE)) {
  data \%>\%
    filter(!is.na(version_part(versions, "dev")))
}

# Filter for major version 1
if (require(dplyr, quietly = TRUE)) {
  data \%>\%
    filter(version_part(versions, "major") == 1)
}
}
