% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versionr-package.R
\docType{package}
\name{versionr-package}
\alias{versionr}
\alias{versionr-package}
\title{versionr: Create, and Manipulate and Sort Version Numbers}
\description{
This package provides an S3 class designed to remove the friction
    of working with version numbers in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wurli/versionr}
  \item \url{https://wurli.github.io/versionr/}
}

}
\author{
\strong{Maintainer}: Jacob Scott \email{jscott2718@gmail.com}

}
\keyword{internal}
