% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versionR.R
\docType{package}
\name{versionR}
\alias{versionR}
\alias{saveVersion}
\title{versionR}
\usage{
saveVersion(x)
}
\arguments{
\item{x}{The name of an R object in quotes}
}
\value{
The R object is saved as a .rds file within a subdirectory of your working directory at ../Env/Versions/.
}
\description{
Helpful R functions for version control of R data objects.
}
\examples{
saveVersion("df")
}
