% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_version_part}
\alias{which_version_part}
\title{Get the index of a version number part}
\usage{
which_version_part(which)
}
\arguments{
\item{which}{Either an option in \code{getOption("versionr.parts")} or a number}
}
\description{
Get the index of a version number part
}
