% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versionr-package.R
\docType{package}
\name{versionr-package}
\alias{versionr}
\alias{versionr-package}
\title{versionr: Create, Use and Manipulate Version Numbers}
\description{
This package provides an S3 class to make working with 
    version numbers much easier.
}
\author{
\strong{Maintainer}: Jacob Scott \email{jscott2718@gmail.com}

}
\keyword{internal}
